pyproject.toml
[tool.poetry]
name = "mtu chapel system"
version = "0.1.0"
description = ""
authors = ["dakohhh <wiizzydreadmill@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Flask = "^3.0.3"
numpy = "^1.26.4"
pandas = "^2.2.2"
python-dateutil = "^2.9.0.post0"
mongoengine = "^0.28.2"
fastapi = "^0.110.1"
bcrypt = "^4.1.2"
pyjwt = "^2.8.0"
beanie = "^1.25.0"
python-multipart = "^0.0.9"
uvicorn = "^0.29.0"
email-validator = "^2.1.1"
pydantic-settings = "^2.2.1"
certifi = "^2024.2.2"
requests = "^2.31.0"
python-docx = "^1.1.0"
openpyxl = "^3.1.4"
humanfriendly = "^10.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
